src/third_party/blink/web_tests/external/wpt/css/selectors/invalidation/is-expected.txt

This is a testharness.js-based test.
PASS Preconditions.
PASS Invalidate :is() for simple selector arguments.
PASS Invalidate :is() for compound selector arguments.
PASS Invalidate :is() for complex selector arguments.
FAIL Invalidate nested :is(). assert_equals: expected "rgb(0, 0, 0)" but got "rgb(255, 0, 0)"
FAIL Test specificity of :is(). assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
